fix: clippy warnings and upgrade pyo3 to 0.24
- Upgrade pyo3 from 0.20.3 to 0.24 (resolves security advisory)
- Fix manual_is_multiple_of lint (use is_multiple_of() instead of % == 0)
- Fix trim_split_whitespace lint (remove redundant trim())
- Fix io_other_error lint (use io::Error::other())
- Fix len_zero lint (use !is_empty() instead of len() > 0)
- Fix useless_vec lint (use array literals where possible)
- Fix manual_range_contains lint (use Range::contains())
- Fix option_map_unit_fn lint (use if let instead of map)
- Fix approx_constant lint (use std::f64::consts::PI)
- Remove unused variables
- Enable criterion benchmarking in Cargo.toml
- Apply cargo fmt formatting

+++ ironbase-core/src/btree.rs
@@ -453,16 +453,12 @@ mod tests {
         // Insert in non-sequential order
         let keys = vec![50, 25, 75, 10, 30, 60, 90, 5, 15, 20];
         for &k in &keys {
-            tree.insert(IndexKey::Int(k), DocumentId::Int(k as i64))
-                .unwrap();
+            tree.insert(IndexKey::Int(k), DocumentId::Int(k)).unwrap();
         }
 
         // All should be searchable
         for &k in &keys {
-            assert_eq!(
-                tree.search(&IndexKey::Int(k)),
-                Some(DocumentId::Int(k as i64))
-            );
+            assert_eq!(tree.search(&IndexKey::Int(k)), Some(DocumentId::Int(k)));
         }
     }
